For testing 0.5.4.6, add support for 4 "CDN" download servers.
This makes using the replacement /serverlist document on by default, and
now there are three registry entries containing two possible default lists
of servers. The thinking here is to move to listing four servers by default
rather than the default two CDN-type servers that seems to be standard for
Valve to return.

To get more download parallelism Valve seem to be forcing the use of "CS"
type servers instead of using more "CDN" type ones for reasons we can't
really speculate about, but since "CS" type servers are identified by a
fixed IP address they're extremely annoying to try and remap. Instead, I
am replacing the /serverlist document to suppress their use (and also have
a URL blocking rule to cause them to be dropped as a backstop) and have
instead have a serverlist that lists all of content1 through content4 as
CDN type servers instead of just the content1 and content3 ones, so that
if more servers are unmetered they'll get rotated through as before but
Steam will be trying 4 in parallel instead of just 2 of them.

In this checkin, the original "generic" serverlist document has been
renamed "standard" and the new longer serverlist is installed under two
names, "generic" and "fourway" with the intent that people can either cut-
and-paste the other values into the "generic" slot instead of writing a
custom rule or they can use custom rules pointing at one of the other lists
explicitly if they want to instead.