Massively cull the list of Internode/iiNet Steam servers as many are gone.
It seems many of the previously unmetered servers used by Internode have
been decommissioned during recent months (and no-one seems to have noticed,
most likely due to Valve using the "CS" server type and evading the non-test
versions of steam-limiter). I've narrowed things down to a smaller list of
servers that definitely work for me, plus steam.cdn.on.net which doesn't
work for me but is still as far as I can find out the main thing that is
supposed to be right for iiNet.