New rewrite system that can modify and pass/fail HTTP requests using host:
I was already parsing outbound HTTP requests to get the host:, but now I
go so far as to look for host-based patterns and allow a host pattern to
both pass/fail an outbound request using the usual convention but also to
allow the host to be replaced (just as the earlier code allowed URLs to
be replaced, although I wasn't using that).

This should allow more flexibility in dealing with redirecting Steam to
ISP-specific servers that have particular requirements, but in the most
usual case this will make it possible to write rules that deal with the
now quite complex /serverlist/ documents that Valve returns for ISPs like
iiNet and Internode, and to allow Steam to use the stock returned document
rather than having to replace it.

By not replacing the document I can get filters to apply instantly on hook
and unhook (without worrying about how the serverlist is cached), but thanks
to a new set of global blacklist rules like
   //*.steampowered.com/=*;//*/depot/=
what happens is that I pass all requests with a host: that refers to one of
the main Steam domains but I reject all other requests that begin with /depot
if they didn't come from something that had a host match. This has the same
effect as replacing the server list in rejecting any additionally supplied
servers, but allows whitelisting of CDN-type servers with custom names in
the rule by adding a pass element like
    //steam.cdn.on.net=*
which allows that host to be allowed through (as it would be supplied by
Valve's serverlist by default anyway for the ISPs that have it unmetered).

I'm still deciding on how to publish the above element and have it picked
up by the configuration script just so as not to have these new-style rules
interfere too much with older steam-limiter versions.

The configuration webservice changes are just something I plan to experiment
with for now, while I publish the 0.6.1.1 exe as a test build for iiNet and
Internode testers to use.