Minor uninstall fix; install by default to %AppData%; new profile UI
This adds support in setfilter.js to set a new host configuration string for
host-level configuration (using the 'allow' string I added to the bundle in
the 0.6.1.1 test version).

There is also now code to read this and to let it be configured through the
profile UI using a new edit box. I've also considerably expanded the size of
the profile edit window and have the option of using multi-line edit controls
for more complex rules (and I have expanded the filter rule parser to take
newlines in rules as equivalent to ';' separators, experimentally for now).

[ I have the option of going back to the old single-line profile edit boxes
  depending on how this is received; the key problem is the handling of the
  return key in multiline edit boxes. Those edit boxes are awful to use if
  you don't enable ES_WANTRETURN. ]

I've also fixed a problem with the uninstalls that was reported after fixing
a subtle Steam crash bug; while investigating that I've moved all the
uninstall data for steam-limiter into HKEY_CURRENT_USER instead of the old
HKEY_LOCAL_MACHINE and made the new install location of steam-limiter be
in the %AppData%\Local profile folder instead of %ProgramFiles% so that
the whole install is now completely per-user instead of being a hybrid of
per-user and per-machine as at present.

[ I may in future re-add the option of installing per-machine, depending
  how people receive the change to being fully a per-user install. ]

It also now correctly handles being installed to custom locations, and
records the custom location so that once installed to a custom location
it reinstalls there in the future.