Make a 0.7.0.1 version that fixes a stupid mistake in the install script.
Basically, if installed in the "Program Files" directory, the script line
that deleted the old per-machine install directory ... deleted the newly
installed version, since I had it in the wrong place. Now it's in a place
where that won't happen, and more importantly I don't try and remove the
old machine-wide install at all if that's where the new install is going.