Retire the rule for Optus, adjust the rule for Telstra
It seems that there's a hidden DNS name for the Telstra server that there's
no way I could have discovered since there's no reverse DNS entry for it,
and I'm making a guess that it's being advertised through the CDN host
system so I can use the 0.7 'allow' rules for it. So, use that DNS name.

But then again, none of the Australian ISPs document what they do accurately,
so I guess complaining about Telstra's inabilty to document their stuff is
hardly news.