Experimental changes to allow revectoring of CS-type servers.
Work on https://code.google.com/p/steam-limiter/issues/detail?id=30

The idea here is that I now allow substitution rules which can return all
kinds of HTTP status codes instead of just 200, as well as inline content
that can substitute for a response (for simple responses) and handling of
POST data in redirections - previously, if we substituted the response to a
POST, we didn't discard the body of the POST.

[ Note that the discard system is not as robust as it could be, in that it
  depends somewhat on the HTTP header and POST body being written in two
  separate send () calls. As it happens Steam does do this, but in principle
  if we wanted to make this 100% robust I'd need to do tricky things to deal
  with the HTTP headers and body going out in a single write. ]

To help avoid problems inside Steam caused by how I was faking out the "CS"
servers, I then use this expanded replacement system to generate a fake
response to the /initsession/ POST and stub out the /authdepot/ POST, so
that neither goes out on the wire.

By doing this, I now have freedom to create a host-type rule that can work
on the CS-server DNS names, and in particular rewrite those as they go out
to point at an ISP-level proxy instead of a real CS server, which should
help with the state of affairs where most unmetering is done through a simple
Squid cache rather than a complex list of IPs.

I'm not sure whether this will stay as 0.7.0.5 or it'll become a 0.7.1.0
after testing, we shall see.