Prepare for a new version 0.7.1.0 with a new download system.
At Jan 15 2014, Google Code is turning off its download facility; this is
a bit awkward since the upgrade script I was using it set to only download
from there, and good alternatives are thin on the ground.

This changeset packages up both the recent 0.7.0.4 and 0.7.0.5 fixes, but
since those also need new rulesets as well as a new download system, I've
done some refactoring of the configuration webservice so that it can handle
pre-0.7.1.0 versions and more recent versions separately and deliver them
different configurations, including different suggested "next" versions and
upgrade paths.

Since my binary packages are so small, I'm going to experiment with the new
home for download packages being static files in the webservice; as well,
the binaries will end up being checked in and thus should still be available
from googlecode.com via the web interface to the version control system.
Neither mechanism is designed for big stuff, but since my binaries are so
small they should be fine, and by having the configurations split I'm going
to be able to experiment with different things.

[ The "ideal" approach for using GAE to serve binary content is through the
  blobstore, and there's separate quota metering for that. I should be miles
  away from any quota issue there, though. ]

I'm just going to start testing the new GAE code on a new version of that
so that I can be reasonably sure I have my ducks in a row before making the
new GAE code the "default" version on steam-limiter.appspot.com