Revise the rules for iPrimus per DH on Whirlpool
Per http://forums.whirlpool.net.au/forum-replies.cfm?t=1482767&p=10#r191
it seems iPrimus just unmeters things that are connected via IX's, so
those are listed. I'll ask DH to test the new rules with v0.7.1.0 but
this should be considerably better for iPrimus than the old rules, and
doesn't need the new features so can be much the same for old and new rule
formats.